lab2 fix
