+++ src/lab1/MagicCube.kt
@@ -17,4 +17,5 @@ class MagicCube {
 fun main() {
     println("В чем вопрос?")
     MagicCube.answer(readLine()!!)
+
 }
