+++ src/lab2/cat_jump/Furniture.kt
@@ -1,0 +1,8 @@
+package lab2.cat_jump
+
+abstract class Furniture(
+    val height: Double,
+    val color: String,
+    val material: String,
+    val length: Double
+)
