+++ src/lab2/cat_jump/Table.kt
@@ -1,0 +1,6 @@
+package lab2.cat_jump
+
+class Table(height: Double, color: String, material: String, length: Double, val numOfLegs: Int) :
+    Furniture(height, color, material, length) {
+
+}
